chore: optimize import,switch,resource release,etc.

+++ minidb-server/src/main/java/com/minidb/server/exec/functions/Kernels.java
@@ -132,14 +132,14 @@ public final class Kernels {
                                           ScalarKernels.LongCompare cmp, SqlKind kind) {
         for (int i = 0; i < l.getValueCount(); i++) {
             if (l.isNull(i) || r.isNull(i)) { out.setNull(i); continue; }
-            out.setSafe(i, compareToBool(cmp.apply((long) l.get(i), r.get(i)), kind) ? 1 : 0);
+            out.setSafe(i, compareToBool(cmp.apply(l.get(i), r.get(i)), kind) ? 1 : 0);
         }
     }
     public static void fillCompareLongInt(BigIntVector l, IntVector r, BitVector out,
                                           ScalarKernels.LongCompare cmp, SqlKind kind) {
         for (int i = 0; i < l.getValueCount(); i++) {
             if (l.isNull(i) || r.isNull(i)) { out.setNull(i); continue; }
-            out.setSafe(i, compareToBool(cmp.apply(l.get(i), (long) r.get(i)), kind) ? 1 : 0);
+            out.setSafe(i, compareToBool(cmp.apply(l.get(i), r.get(i)), kind) ? 1 : 0);
         }
     }
